requirements.txt
fire~=0.5.0
torch~=1.10.2+cpu
numpy~=1.19.5
scikit-learn~=0.24.2
matplotlib~=3.3.4
torchvision~=0.11.3+cpu
tqdm~=4.64.0
Pillow~=8.3.2
seaborn~=0.11.2